+++ src/main/java/org/guanzon/cas/inventory/models/Inv_Master.java
@@ -1,0 +1,4 @@
+package org.guanzon.cas.inventory.models;
+
+public class Inv_Master {
+}
