+++ src/main/java/org/guanzon/cas/inventory/models/Inventory.java
@@ -1,0 +1,4 @@
+package org.guanzon.cas.inventory.models;
+
+public class Inventory {
+}
