+++ src/main/java/org/guanzon/cas/inventory/models/Model_Inv_Stock_Request_Detail.java
@@ -1,7 +1,7 @@
 
 
 
-package org.guanzon.cas.model.inventory;
+package org.guanzon.cas.inventory.models;
 
 import java.lang.reflect.Method;
 import java.math.BigDecimal;
