+++ src/main/java/org/guanzon/cas/inventory/models/Model_Inv_Stock_Request_Master.java
@@ -1,10 +1,7 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package org.guanzon.cas.model.inventory;
+package org.guanzon.cas.inventory.models;
 
 import java.lang.reflect.Method;
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Types;
@@ -13,60 +10,57 @@ import javax.sql.rowset.CachedRowSet;
 import org.guanzon.appdriver.base.GRider;
 import org.guanzon.appdriver.base.MiscUtil;
 import org.guanzon.appdriver.base.SQLUtil;
-import org.guanzon.appdriver.constant.EditMode;
 import org.guanzon.appdriver.constant.RecordStatus;
+import org.guanzon.appdriver.constant.EditMode;
 import org.guanzon.appdriver.iface.GEntity;
 import org.json.simple.JSONObject;
 
+
 /**
- *
- * @author User
+ * @author Michael Cuison
  */
-public class Model_Inv_Hist_Ledger implements GEntity {
-
-    final String XML = "Model_Inv_Hist_Ledger.xml";
-
+public class Model_Inv_Stock_Request_Master implements GEntity{
+    final String XML = "Model_Inv_Stock_Request_Master.xml";
+    
     GRider poGRider;                //application driver
     CachedRowSet poEntity;          //rowset
     JSONObject poJSON;              //json container
     int pnEditMode;                 //edit mode
-
+    
     /**
      * Entity constructor
-     *
+     * 
      * @param foValue - GhostRider Application Driver
      */
-    public Model_Inv_Hist_Ledger(GRider foValue) {
-        if (foValue == null) {
+    public Model_Inv_Stock_Request_Master(GRider foValue){
+        if (foValue == null){
             System.err.println("Application Driver is not set.");
             System.exit(1);
         }
-
+        
         poGRider = foValue;
-
+        
         initialize();
     }
-
+    
     /**
      * Gets edit mode of the record
-     *
      * @return edit mode
      */
     @Override
     public int getEditMode() {
         return pnEditMode;
     }
-
+    
     /**
      * Gets the column index name.
-     *
      * @param fnValue - column index number
      * @return column index name
      */
     @Override
     public String getColumn(int fnValue) {
         try {
-            return poEntity.getMetaData().getColumnLabel(fnValue);
+            return poEntity.getMetaData().getColumnLabel(fnValue); 
         } catch (SQLException e) {
         }
         return "";
@@ -74,7 +68,6 @@ public class Model_Inv_Hist_Ledger implements GEntity {
 
     /**
      * Gets the column index number.
-     *
      * @param fsValue - column index name
      * @return column index number
      */
@@ -90,33 +83,31 @@ public class Model_Inv_Hist_Ledger implements GEntity {
 
     /**
      * Gets the total number of column.
-     *
      * @return total number of column
      */
     @Override
     public int getColumnCount() {
         try {
-            return poEntity.getMetaData().getColumnCount();
+            return poEntity.getMetaData().getColumnCount(); 
         } catch (SQLException e) {
             e.printStackTrace();
         }
-
+        
         return -1;
     }
 
     /**
      * Gets the table name.
-     *
      * @return table name
      */
     @Override
     public String getTable() {
-        return "Inv_Hist_Ledger";
+        return "Made";
     }
-
+    
     /**
      * Gets the value of a column index number.
-     *
+     * 
      * @param fnColumn - column index number
      * @return object value
      */
@@ -132,7 +123,7 @@ public class Model_Inv_Hist_Ledger implements GEntity {
 
     /**
      * Gets the value of a column index name.
-     *
+     * 
      * @param fsColumn - column index name
      * @return object value
      */
@@ -145,25 +136,23 @@ public class Model_Inv_Hist_Ledger implements GEntity {
         }
         return null;
     }
-
+    
     /**
      * Sets column value.
-     *
+     * 
      * @param fnColumn - column index number
      * @param foValue - value
      * @return result as success/failed
      */
     @Override
     public JSONObject setValue(int fnColumn, Object foValue) {
-        try {
+        try {  
             poJSON = MiscUtil.validateColumnValue(System.getProperty("sys.default.path.metadata") + XML, MiscUtil.getColumnLabel(poEntity, fnColumn), foValue);
-            if ("error".equals((String) poJSON.get("result"))) {
-                return poJSON;
-            }
-
+            if ("error".equals((String) poJSON.get("result"))) return poJSON;
+            
             poEntity.updateObject(fnColumn, foValue);
             poEntity.updateRow();
-
+            
             poJSON = new JSONObject();
             poJSON.put("result", "success");
             poJSON.put("value", getValue(fnColumn));
@@ -172,13 +161,13 @@ public class Model_Inv_Hist_Ledger implements GEntity {
             poJSON.put("result", "error");
             poJSON.put("message", e.getMessage());
         }
-
+        
         return poJSON;
     }
 
     /**
      * Sets column value.
-     *
+     * 
      * @param fsColumn - column index name
      * @param foValue - value
      * @return result as success/failed
@@ -186,7 +175,7 @@ public class Model_Inv_Hist_Ledger implements GEntity {
     @Override
     public JSONObject setValue(String fsColumn, Object foValue) {
         poJSON = new JSONObject();
-
+        
         try {
             return setValue(MiscUtil.getColumnIndex(poEntity, fsColumn), foValue);
         } catch (SQLException e) {
@@ -196,18 +185,19 @@ public class Model_Inv_Hist_Ledger implements GEntity {
         }
         return poJSON;
     }
-
+    
     /**
      * Set the edit mode of the entity to new.
-     *
+     * 
      * @return result as success/failed
      */
     @Override
     public JSONObject newRecord() {
         pnEditMode = EditMode.ADDNEW;
-
+        
         //replace with the primary key column info
-//        setStockID(MiscUtil.getNextCode(getTable(), "sStockIDx", true, poGRider.getConnection(), ""));
+        setTransactionNumber(MiscUtil.getNextCode(getTable(), "sMadeIDxx", true, poGRider.getConnection(), poGRider.getBranchCode()));
+        
         poJSON = new JSONObject();
         poJSON.put("result", "success");
         return poJSON;
@@ -215,29 +205,29 @@ public class Model_Inv_Hist_Ledger implements GEntity {
 
     /**
      * Opens a record.
-     *
+     * 
      * @param fsCondition - filter values
      * @return result as success/failed
      */
     @Override
     public JSONObject openRecord(String fsCondition) {
         poJSON = new JSONObject();
-
-        String lsSQL = getSQL();
-
+        
+        String lsSQL = MiscUtil.makeSelect(this, "xCategrNm");
+        
         //replace the condition based on the primary key column of the record
-        lsSQL = MiscUtil.addCondition(lsSQL, "a.sStockIDx = " + SQLUtil.toSQL(fsCondition));
-
+        lsSQL = MiscUtil.addCondition(lsSQL, fsCondition);
+        
         ResultSet loRS = poGRider.executeQuery(lsSQL);
-
+        
         try {
-            if (loRS.next()) {
-                for (int lnCtr = 1; lnCtr <= loRS.getMetaData().getColumnCount(); lnCtr++) {
+            if (loRS.next()){
+                for (int lnCtr = 1; lnCtr <= loRS.getMetaData().getColumnCount(); lnCtr++){
                     setValue(lnCtr, loRS.getObject(lnCtr));
                 }
-
+                
                 pnEditMode = EditMode.UPDATE;
-
+                
                 poJSON.put("result", "success");
                 poJSON.put("message", "Record loaded successfully.");
             } else {
@@ -248,30 +238,29 @@ public class Model_Inv_Hist_Ledger implements GEntity {
             poJSON.put("result", "error");
             poJSON.put("message", e.getMessage());
         }
-
+        
         return poJSON;
     }
 
     /**
      * Save the entity.
-     *
+     * 
      * @return result as success/failed
      */
     @Override
     public JSONObject saveRecord() {
         poJSON = new JSONObject();
-
-        if (pnEditMode == EditMode.ADDNEW || pnEditMode == EditMode.UPDATE) {
+        
+        if (pnEditMode == EditMode.ADDNEW || pnEditMode == EditMode.UPDATE){
             String lsSQL;
-            if (pnEditMode == EditMode.ADDNEW) {
+            if (pnEditMode == EditMode.ADDNEW){
                 //replace with the primary key column info
-//                setStockID(MiscUtil.getNextCode(getTable(), "sStockIDx", false, poGRider.getConnection(), ""));
-
-                setModifiedDate(poGRider.getServerDate());
+                setTransactionNumber(MiscUtil.getNextCode(getTable(), "sMadeIDxx", true, poGRider.getConnection(), poGRider.getBranchCode()));
+                
                 lsSQL = makeSQL();
-
-                if (!lsSQL.isEmpty()) {
-                    if (poGRider.executeQuery(lsSQL, getTable(), poGRider.getBranchCode(), "") > 0) {
+                
+                if (!lsSQL.isEmpty()){
+                    if (poGRider.executeQuery(lsSQL, getTable(), poGRider.getBranchCode(), "") > 0){
                         poJSON.put("result", "success");
                         poJSON.put("message", "Record saved successfully.");
                     } else {
@@ -283,26 +272,25 @@ public class Model_Inv_Hist_Ledger implements GEntity {
                     poJSON.put("message", "No record to save.");
                 }
             } else {
-                Model_Inv_Hist_Ledger loOldEntity = new Model_Inv_Hist_Ledger(poGRider);
-
+                Model_Inv_Stock_Request_Master loOldEntity = new Model_Inv_Stock_Request_Master(poGRider);
+                
                 //replace with the primary key column info
-                JSONObject loJSON = loOldEntity.openRecord(this.getStockID());
-
-                setModifiedDate(poGRider.getServerDate());
-                if ("success".equals((String) loJSON.get("result"))) {
+                JSONObject loJSON = loOldEntity.openRecord(this.getTransactionNumber());
+                
+                if ("success".equals((String) loJSON.get("result"))){
                     //replace the condition based on the primary key column of the record
-                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sStockIDx = " + SQLUtil.toSQL(this.getStockID()),  "xBarCodex»xDescript»xWHouseNm»nQtyOnHnd");
-                    if (!lsSQL.isEmpty()) {
-                        if (poGRider.executeQuery(lsSQL, getTable(), poGRider.getBranchCode(), "") > 0) {
+                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sMadeIDxx = " + SQLUtil.toSQL(this.getTransactionNumber()), "xCategrNm");
+                    
+                    if (!lsSQL.isEmpty()){
+                        if (poGRider.executeQuery(lsSQL, getTable(), poGRider.getBranchCode(), "") > 0){
                             poJSON.put("result", "success");
                             poJSON.put("message", "Record saved successfully.");
                         } else {
                             poJSON.put("result", "error");
                             poJSON.put("message", poGRider.getErrMsg());
-                        }   
+                        }
                     } else {
                         poJSON.put("result", "success");
-                        poJSON.put("continue", true);
                         poJSON.put("message", "No updates has been made.");
                     }
                 } else {
@@ -315,10 +303,10 @@ public class Model_Inv_Hist_Ledger implements GEntity {
             poJSON.put("message", "Invalid update mode. Unable to save record.");
             return poJSON;
         }
-
+        
         return poJSON;
     }
-
+    
     /**
      * Prints all the public methods used<br>
      * and prints the column names of this entity.
@@ -326,502 +314,421 @@ public class Model_Inv_Hist_Ledger implements GEntity {
     @Override
     public void list() {
         Method[] methods = this.getClass().getMethods();
-
+        
         System.out.println("--------------------------------------------------------------------");
         System.out.println("LIST OF PUBLIC METHODS FOR " + this.getClass().getName() + ":");
         System.out.println("--------------------------------------------------------------------");
         for (Method method : methods) {
             System.out.println(method.getName());
         }
-
+        
         try {
             int lnRow = poEntity.getMetaData().getColumnCount();
-
+        
             System.out.println("--------------------------------------------------------------------");
             System.out.println("ENTITY COLUMN INFO");
             System.out.println("--------------------------------------------------------------------");
             System.out.println("Total number of columns: " + lnRow);
             System.out.println("--------------------------------------------------------------------");
 
-            for (int lnCtr = 1; lnCtr <= lnRow; lnCtr++) {
+            for (int lnCtr = 1; lnCtr <= lnRow; lnCtr++){
                 System.out.println("Column index: " + (lnCtr) + " --> Label: " + poEntity.getMetaData().getColumnLabel(lnCtr));
-                if (poEntity.getMetaData().getColumnType(lnCtr) == Types.CHAR
-                        || poEntity.getMetaData().getColumnType(lnCtr) == Types.VARCHAR) {
+                if (poEntity.getMetaData().getColumnType(lnCtr) == Types.CHAR ||
+                    poEntity.getMetaData().getColumnType(lnCtr) == Types.VARCHAR){
 
                     System.out.println("Column index: " + (lnCtr) + " --> Size: " + poEntity.getMetaData().getColumnDisplaySize(lnCtr));
                 }
             }
         } catch (SQLException e) {
         }
-
-    }
-
-    /**
-     * Sets the sStockIDx of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setStockID(String fsValue) {
-        return setValue("sStockIDx", fsValue);
-    }
-
-    /**
-     * @return The sStockIDx of this record.
-     */
-    public String getStockID() {
-        return (String) getValue("sStockIDx");
-    }
-
-    /**
-     * @return The sBranchCd of this record.
-     */
-    public String getBranchCode() {
-        return (String) getValue("sBranchCd");
+        
     }
     
     /**
-     * Sets the sBrandCde of this record.
-     *
-     * @param fsValue
+     * Sets the TransactionNumber of this record.
+     * 
+     * @param fsValue 
      * @return result as success/failed
      */
-    public JSONObject setBranchCode(String fsValue) {
-        return setValue("sBranchCd", fsValue);
-    }
-    
-  
-    /**
-     * @return The sWHouseID of this record.
-     */
-    public String getWHouseID() {
-        return (String) getValue("sWHouseID");
+    public JSONObject setTransactionNumber(String fsValue){
+        return setValue("sTransNox", fsValue);
     }
     
     /**
-     * Sets the sWHouseID of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
+     * @return The TransactionNumber of this record.
      */
-    public JSONObject setWHouseID(String fsValue) {
-        return setValue("sWHouseID", fsValue);
+    public String getTransactionNumber(){
+        return (String) getValue("sTransNox");
     }
     
     /**
-     * @return The nLedgerNo of this record.
+     * Sets the BranchCode of this record.
+     * 
+     * @param fnValue 
+     * @return result as success/failed
      */
-    public int getLedgerNo() {
-        return (int) getValue("nLedgerNo");
+    public JSONObject setBranchCode(Integer fnValue){
+        return setValue("sBranchCd", fnValue);
     }
     
     /**
-     * Sets the nLedgerNo of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
+     * @return The BranchCode of this record. 
      */
-    public JSONObject setLedgerNo(Number fsValue) {
-        return setValue("nLedgerNo", fsValue);
+    public Integer getBranchCode(){
+        return (Integer) getValue("sBranchCd");
     }
-
-    /**
-     * Sets the date and time the record was transact.
-     *
-     * @param fdValue
+    
+     /**
+     * Sets the CategoryCode of this record.
+     * 
+     * @param fsValue 
      * @return result as success/failed
      */
-    public JSONObject setTransactDate(Date fdValue) {
+    public JSONObject setCategoryCode(String fsValue){
+        return setValue("sCategrCd", fsValue);
+    }
+    
+    /**
+     * @return The CategoryCode of this record. 
+     */
+    public String getCategoryCode(){
+        return (String) getValue("sCategrCd");
+    }
+    
+     /**
+     * Sets the Transaction of this record.
+     * 
+     * @param fdValue 
+     * @return result as success/failed
+     */
+    public JSONObject setTransaction(Date fdValue){
         return setValue("dTransact", fdValue);
     }
-
+    
     /**
-     * @return The date and time the record was transact.
+     * @return The Transaction of this record. 
      */
-    public Date getTransactDate() {
+    public Date getTransaction(){
         return (Date) getValue("dTransact");
     }
-
     
-    /**
-     * @return The sSourceCd of this record.
+     /**
+     * Sets the ReferenceNumber of this record.
+     * 
+     * @param fsValue 
+     * @return result as success/failed
      */
-    public String getSourceCode() {
-        return (String) getValue("sSourceCd");
+    public JSONObject setReferenceNumber(String fsValue){
+        return setValue("sReferNox", fsValue);
     }
     
     /**
-     * Sets the sSourceCd of this record.
-     *
-     * @param fsValue
+     * @return The ReferenceNumber of this record. 
+     */
+    public String getReferenceNumber(){
+        return (String) getValue("sReferNox");
+    }
+    
+    /**
+     * Sets the Remarks of this record.
+     * 
+     * @param fsValue 
      * @return result as success/failed
      */
-    public JSONObject setSourceCode(String fsValue) {
+    public JSONObject setRemarks(String fsValue){
+        return setValue("sRemarksx", fsValue);
+    }
+    
+    /**
+     * @return The Remarks of this record. 
+     */
+    public String getRemarks(){
+        return (String) getValue("sRemarksx");
+    }
+    
+    /**
+     * Sets the IssNotes of this record.
+     * 
+     * @param fsValue 
+     * @return result as success/failed
+     */
+    public JSONObject setIssNotes(String fsValue){
+        return setValue("sIssNotes", fsValue);
+    }
+    
+    /**
+     * @return The IssNotes of this record. 
+     */
+    public String getIssNotes(){
+        return (String) getValue("sIssNotes");
+    }
+    
+    /**
+     * Sets the CurrentInventory of this record.
+     * 
+     * @param fnValue 
+     * @return result as success/failed
+     */
+    public JSONObject setCurrentInventory(Integer fnValue){
+        return setValue("nCurrInvx", fnValue);
+    }
+    
+    /**
+     * @return The CurrentInventory of this record. 
+     */
+    public Integer getCurrentInventory(){
+        return (Integer) getValue("nCurrInvx");
+    }
+    
+    /**
+     * Sets the EstimatedInventory of this record.
+     * 
+     * @param fnValue 
+     * @return result as success/failed
+     */
+    public JSONObject setEstimatedInventory(Integer fnValue){
+        return setValue("nEstInvxx", fnValue);
+    }
+    
+    /**
+     * @return The EstimatedInventory of this record. 
+     */
+    public Integer getEstimatedInventory(){
+        return (Integer) getValue("nEstInvxx");
+    }
+    
+    /**
+     * Sets the Approved of this record.
+     * 
+     * @param fsValue 
+     * @return result as success/failed
+     */
+    public JSONObject setApproved(String fsValue){
+        return setValue("sApproved", fsValue);
+    }
+    
+    /**
+     * @return The Approved of this record. 
+     */
+    public String getApproved(){
+        return (String) getValue("sApproved");
+    }
+    
+    /**
+     * Sets the Approved of this record.
+     * 
+     * @param fdValue 
+     * @return result as success/failed
+     */
+    public JSONObject setApprovedDate(Date fdValue){
+        return setValue("dApproved", fdValue);
+    }
+    
+    /**
+     * @return The Approved of this record. 
+     */
+    public Date getApprovedDate(){
+        return (Date) getValue("dApproved");
+    }
+    
+    /**
+     * Sets the ApproveCode of this record.
+     * 
+     * @param fsValue 
+     * @return result as success/failed
+     */
+    public JSONObject setApproveCode(String fsValue){
+        return setValue("sAprvCode", fsValue);
+    }
+    
+    /**
+     * @return The ApproveCode of this record. 
+     */
+    public String getApproveCode(){
+        return (String) getValue("sAprvCode");
+    }
+    
+    /**
+     * Sets the EntryNumber of this record.
+     * 
+     * @param fnValue 
+     * @return result as success/failed
+     */
+    public JSONObject setEntryNumber(Integer fnValue){
+        return setValue("nEntryNox", fnValue);
+    }
+    
+    /**
+     * @return The EntryNumber of this record. 
+     */
+    public Integer getEntryNumber(){
+        return (Integer) getValue("nEntryNox");
+    }
+    
+    /**
+     * Sets the SourceCode of this record.
+     * 
+     * @param fsValue 
+     * @return result as success/failed
+     */
+    public JSONObject setSourceCode(String fsValue){
         return setValue("sSourceCd", fsValue);
     }
     
     /**
-     * @return The sSourceNo of this record.
+     * @return The SourceCode of this record. 
      */
-    public String getSourceNo() {
+    public String getSourceCode(){
+        return (String) getValue("sSourceCd");
+    }
+    
+    /**
+     * Sets the SourceNumber of this record.
+     * 
+     * @param fsValue 
+     * @return result as success/failed
+     */
+    public JSONObject setSourceNumber(String fsValue){
+        return setValue("sSourceNo", fsValue);
+    }
+    
+    /**
+     * @return The SourceNumber of this record. 
+     */
+    public String getSourceNumber(){
         return (String) getValue("sSourceNo");
     }
     
     /**
-     * Sets the sSourceNo of this record.
-     *
-     * @param fsValue
+     * Sets the Confirm of this record.
+     * 
+     * @param fsValue 
      * @return result as success/failed
      */
-    public JSONObject setSourceNo(String fsValue) {
-        return setValue("sSourceNo", fsValue);
+    public JSONObject setConfirm(String fsValue){
+        return setValue("cConfirmd", fsValue);
     }
-
+    
     /**
+     * @return The Confirm of this record. 
+     */
+    public String getConfirm(){
+        return (String) getValue("cConfirmd");
+    }
+    
     /**
-     * Sets the nQtyInxxx .
+     * Sets the TransactionStatus of this record.
      * 
      * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityIn(Number fsValue){
-        return setValue("nQtyInxxx", fsValue);
-    }
-    
-    /**
-     * @return The nQtyInxxx. 
-     */
-    public int getQuantityIn(){
-        return (int) getValue("nQtyInxxx");
-    }
-    
-    /**
-    /**
-     * Sets the nQtyOutxx .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityOut(Number fsValue){
-        return setValue("nQtyOutxx", fsValue);
-    }
-    
-    /**
-     * @return The nQtyOutxx. 
-     */
-    public int getQuantityOut(){
-        return (int) getValue("nQtyOutxx");
-    }
-    
-
-    /**
-    /**
-     * Sets the nQtyOrder .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityOrder(Number fsValue){
-        return setValue("nQtyOrder", fsValue);
-    }
-    
-    /**
-     * @return The nQtyOrder. 
-     */
-    public int getQuantityOrder(){
-        return (int) getValue("nQtyOrder");
-    }
-
-    /**
-    /**
-     * Sets the nQtyIssue .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityIssue(Number fsValue){
-        return setValue("nQtyIssue", fsValue);
-    }
-    
-    /**
-     * @return The nQtyIssue. 
-     */
-    public int getQuantityIssue(){
-        return (int) getValue("nQtyIssue");
-    }
-    
-    /**
-    /**
-     * Sets the nPurPrice .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setPurchasePrice(Number fsValue){
-        return setValue("nPurPrice", fsValue);
-    }
-    
-    /**
-     * @return The nUnitPrce. 
-     */
-    public Object getPurchasePrice(){
-        return (Object) getValue("nPurPrice");
-    }
-    
-    /**
-    /**
-     * Sets the nUnitPrce .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setUnitPrice(Number fsValue){
-        return setValue("nUnitPrce", fsValue);
-    }
-    
-    /**
-     * @return The nUnitPrce. 
-     */
-    public Object getUnitPrice(){
-        return (Object) getValue("nUnitPrce");
-    }
-    
-    
-    /**
-     * Sets the nSelPrice .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setSelPrice(Number fsValue){
-        return setValue("nSelPrice", fsValue);
-    }
-    
-    /**
-     * @return The nSelPrice. 
-     */
-    public Double getSelPrice(){
-        return (Double) getValue("nSelPrice");
-    }
-    
-    
-    /**
-    /**
-     * Sets the nQtyOnHnd .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityOnHand(Number fsValue){
-        return setValue("nQtyOnHnd", fsValue);
-    }
-    
-    /**
-     * @return The nQtyOnHnd. 
-     */
-    public int getQuantityOnHand(){
-        return (int) getValue("nQtyOnHnd");
-    }
-    
-    
-    /**
-     * Sets the date and time the record expiration date.
-     *
-     * @param fdValue
      * @return result as success/failed
      */
-    public JSONObject setExpiryDate(Date fdValue) {
-        return setValue("dExpiryxx", fdValue);
-    }
-
-    /**
-     * @return The date and time the record expiration date.
-     */
-    public Date getExpiryDate() {
-        return (Date) getValue("dExpiryxx");
-    } 
-    
-    /**
-     * Sets the xBarCodex of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setBarcode(String fsValue) {
-        return setValue("xBarCodex", fsValue);
-    }
-
-    /**
-     * @return The xBarCodex of this record.
-     */
-    public String getBarcode() {
-        return (String) getValue("xBarCodex");
+    public JSONObject setTransactionStatus(String fsValue){
+        return setValue("cTranStat", fsValue);
     }
     
     /**
-     * Sets the xDescript of this record.
-     *
-     * @param fsValue
+     * @return The TransactionStatus of this record. 
+     */
+    public String getTransactionStatus(){
+        return (String) getValue("cTranStat");
+    }
+    
+    /**
+     * Sets the RecordStatus of this record.
+     * 
+     * @param fsValue 
      * @return result as success/failed
      */
-    public JSONObject setDescript(String fsValue) {
-        return setValue("xDescript", fsValue);
-    }
-
-    /**
-     * @return The xDescript of this record.
-     */
-    public String getDescript() {
-        return (String) getValue("xDescript");
-    }
-    /**
-     * Sets the xWHouseNm of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setWareHouseName(String fsValue) {
-        return setValue("xWHouseNm", fsValue);
-    }
-
-    /**
-     * @return The xWHouseNm of this record.
-     */
-    public String getWareHouseName() {
-        return (String) getValue("xWHouseNm");
-    }
-
-    /**
-     * Sets the Inventory RecdStat of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setRecdStat(String fsValue) {
+    public JSONObject setRecordStatus(String fsValue){
         return setValue("cRecdStat", fsValue);
     }
-
+    
     /**
-     * @return The Category RecdStat of this record.
+     * @return The RecordStatus of this record. 
      */
-    public String getRecdStat() {
+    public String getRecordStatus(){
         return (String) getValue("cRecdStat");
     }
-
+    
     /**
      * Sets record as active.
-     *
+     * 
      * @param fbValue
      * @return result as success/failed
      */
-    public JSONObject setActive(boolean fbValue) {
+    public JSONObject setActive(boolean fbValue){
         return setValue("cRecdStat", fbValue ? "1" : "0");
     }
-
+    
     /**
-     * @return If record is active.
+     * @return If record is active. 
      */
-    public boolean isActive() {
+    public boolean isActive(){
         return ((String) getValue("cRecdStat")).equals("1");
     }
-
+    
     /**
      * Sets the user encoded/updated the record.
-     *
-     * @param fsValue
+     * 
+     * @param fsValue 
      * @return result as success/failed
      */
-    public JSONObject setModifiedBy(String fsValue) {
+    public JSONObject setModifiedBy(String fsValue){
         return setValue("sModified", fsValue);
     }
-
+    
     /**
-     * @return The user encoded/updated the record
+     * @return The user encoded/updated the record 
      */
-    public String getModifiedBy() {
+    public String getModifiedBy(){
         return (String) getValue("sModified");
     }
-
+    
     /**
      * Sets the date and time the record was modified.
-     *
-     * @param fdValue
+     * 
+     * @param fdValue 
      * @return result as success/failed
      */
-    public JSONObject setModifiedDate(Date fdValue) {
+    public JSONObject setModifiedDate(Date fdValue){
         return setValue("dModified", fdValue);
     }
-
+    
     /**
      * @return The date and time the record was modified.
      */
-    public Date getModifiedDate() {
+    public Date getModifiedDate(){
         return (Date) getValue("dModified");
     }
-
+    
     /**
      * Gets the SQL statement for this entity.
-     *
+     * 
      * @return SQL Statement
      */
-    public String makeSQL() {
-        return MiscUtil.makeSQL(this, "xBarCodex»xDescript»xWHouseNm»nQtyOnHnd");
-    }
-
-    /**
-     * Gets the SQL Select statement for this entity.
-     *
-     * @return SelectSQL Statement
-     */
-    public String makeSelectSQL() {
-        return MiscUtil.makeSelect(this, "xBarCodex»xDescript»xWHouseNm»nQtyOnHnd");
+    public String makeSQL(){
+        return MiscUtil.makeSQL(this, "xCategrNm");
     }
     
-    public String getSQL(){
-        return "SELECT" +
-                        "   a.sStockIDx" +
-                        " , a.sBranchCd" +
-                        " , a.sWHouseID" +
-                        " , a.nLedgerNo" +
-                        " , a.dTransact" +
-                        " , a.sSourceCd" +
-                        " , a.sSourceNo" +
-                        " , a.nQtyInxxx" +
-                        " , a.nQtyOutxx" +
-                        " , a.nQtyOrder" +
-                        " , a.nQtyIssue" +
-                        " , a.nPurPrice" +
-                        " , a.nUnitPrce" +
-//                        " , a.nQtyOnHnd" +
-                        " , a.dExpiryxx" +
-                        " , a.sModified" +
-                        " , a.dModified" +
-                        " , b.sBarCodex xBarCodex" +
-                        " , b.sDescript xDescript" +
-                        " , c.sWHouseNm xWHouseNm " +
-                        "FROM Inv_Hist_Ledger a" +
-                        "    LEFT JOIN Inventory b ON a.sStockIDx = b.sStockIDx" +
-                        "    LEFT JOIN Warehouse c ON a.sWhouseID = c.sWhouseID";
-    }
-    private void initialize() {
+    private void initialize(){
         try {
             poEntity = MiscUtil.xml2ResultSet(System.getProperty("sys.default.path.metadata") + XML, getTable());
-
+            
             poEntity.last();
             poEntity.moveToInsertRow();
 
-            MiscUtil.initRowSet(poEntity);
-//            poEntity.updateString("cRecdStat", RecordStatus.ACTIVE);
-
+            MiscUtil.initRowSet(poEntity);      
+            poEntity.updateString("cRecdStat", RecordStatus.ACTIVE);
+            
             poEntity.insertRow();
             poEntity.moveToCurrentRow();
 
             poEntity.absolute(1);
-
+            
             pnEditMode = EditMode.UNKNOWN;
         } catch (SQLException e) {
             e.printStackTrace();
             System.exit(1);
         }
-    }
+    } 
 }
