+++ src/main/java/org/guanzon/cas/inventory/models/Model_Inventory_Sub_Unit.java
@@ -2,9 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package org.guanzon.cas.model.inventory;
+package org.guanzon.cas.inventory.models;
 
 import java.lang.reflect.Method;
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Types;
@@ -22,9 +23,9 @@ import org.json.simple.JSONObject;
  *
  * @author User
  */
-public class Model_Inv_Ledger implements GEntity {
+public class Model_Inventory_Sub_Unit implements GEntity {
 
-    final String XML = "Model_Inv_Ledger.xml";
+    final String XML = "Model_Inventory_Sub_Unit.xml";
 
     GRider poGRider;                //application driver
     CachedRowSet poEntity;          //rowset
@@ -36,7 +37,7 @@ public class Model_Inv_Ledger implements GEntity {
      *
      * @param foValue - GhostRider Application Driver
      */
-    public Model_Inv_Ledger(GRider foValue) {
+    public Model_Inventory_Sub_Unit(GRider foValue) {
         if (foValue == null) {
             System.err.println("Application Driver is not set.");
             System.exit(1);
@@ -55,6 +56,10 @@ public class Model_Inv_Ledger implements GEntity {
     @Override
     public int getEditMode() {
         return pnEditMode;
+    }
+    
+    public void setEditMode(int val) {
+        pnEditMode = val;
     }
 
     /**
@@ -82,7 +87,7 @@ public class Model_Inv_Ledger implements GEntity {
     public int getColumn(String fsValue) {
         try {
             return MiscUtil.getColumnIndex(poEntity, fsValue);
-        } catch (SQLException e) {  
+        } catch (SQLException e) {
             e.printStackTrace();
         }
         return -1;
@@ -111,7 +116,7 @@ public class Model_Inv_Ledger implements GEntity {
      */
     @Override
     public String getTable() {
-        return "Inv_Ledger";
+        return "Inventory_Sub_Unit";
     }
 
     /**
@@ -155,24 +160,24 @@ public class Model_Inv_Ledger implements GEntity {
      */
     @Override
     public JSONObject setValue(int fnColumn, Object foValue) {
-        try {
+        try {              
             poJSON = MiscUtil.validateColumnValue(System.getProperty("sys.default.path.metadata") + XML, MiscUtil.getColumnLabel(poEntity, fnColumn), foValue);
-            if ("error".equals((String) poJSON.get("result"))) {
-                return poJSON;
-            }
-
+            if ("error".equals((String) poJSON.get("result"))) return poJSON;
+            
             poEntity.updateObject(fnColumn, foValue);
             poEntity.updateRow();
-
+            
             poJSON = new JSONObject();
             poJSON.put("result", "success");
             poJSON.put("value", getValue(fnColumn));
+            System.out.println("poJSON = " + poJSON);
+            
         } catch (SQLException e) {
             e.printStackTrace();
             poJSON.put("result", "error");
             poJSON.put("message", e.getMessage());
         }
-
+        
         return poJSON;
     }
 
@@ -184,17 +189,18 @@ public class Model_Inv_Ledger implements GEntity {
      * @return result as success/failed
      */
     @Override
-    public JSONObject setValue(String fsColumn, Object foValue) {
-        poJSON = new JSONObject();
-
+    public JSONObject setValue(String string, Object foValue) {
         try {
-            return setValue(MiscUtil.getColumnIndex(poEntity, fsColumn), foValue);
-        } catch (SQLException e) {
-            e.printStackTrace();
+            return setValue(MiscUtil.getColumnIndex(poEntity, string), foValue);
+        } catch (SQLException ex) {
+            
+            poJSON = new JSONObject();
             poJSON.put("result", "error");
-            poJSON.put("message", e.getMessage());
+            poJSON.put("message", ex.getMessage());
+            return poJSON;
+            
         }
-        return poJSON;
+        
     }
 
     /**
@@ -207,7 +213,7 @@ public class Model_Inv_Ledger implements GEntity {
         pnEditMode = EditMode.ADDNEW;
 
         //replace with the primary key column info
-//        setStockID(MiscUtil.getNextCode(getTable(), "sStockIDx", false, poGRider.getConnection(), ""));
+//        setStockID(MiscUtil.getNextCode(getTable(), "sStockIDx", true, poGRider.getConnection(), poGRider.getBranchCode()));
         poJSON = new JSONObject();
         poJSON.put("result", "success");
         return poJSON;
@@ -226,17 +232,18 @@ public class Model_Inv_Ledger implements GEntity {
         String lsSQL = getSQL();
 
         //replace the condition based on the primary key column of the record
-        lsSQL = MiscUtil.addCondition(lsSQL, "a.sStockIDx = " + SQLUtil.toSQL(fsCondition));
-
+        lsSQL = MiscUtil.addCondition(lsSQL, "a.sItmSubID = " + SQLUtil.toSQL(fsCondition));
+        System.out.print("this is lsSQL openrec == " + lsSQL + "\n");
         ResultSet loRS = poGRider.executeQuery(lsSQL);
-
         try {
             if (loRS.next()) {
                 for (int lnCtr = 1; lnCtr <= loRS.getMetaData().getColumnCount(); lnCtr++) {
                     setValue(lnCtr, loRS.getObject(lnCtr));
                 }
-                pnEditMode = EditMode.UPDATE;
 
+                pnEditMode = EditMode.UPDATE;
+                
+                
                 poJSON.put("result", "success");
                 poJSON.put("message", "Record loaded successfully.");
             } else {
@@ -251,18 +258,22 @@ public class Model_Inv_Ledger implements GEntity {
         return poJSON;
     }
     
-    public JSONObject openRecord(String lsFilter, String fsCondition) {
+    /**
+     * Opens a record.
+     *
+     * @param fsCondition - filter values
+     * @return result as success/failed
+     */
+    public JSONObject openRecordWithCondition(String fsCondition, int fnEntryNox) {
         poJSON = new JSONObject();
 
         String lsSQL = getSQL();
+
         //replace the condition based on the primary key column of the record
-        lsSQL = MiscUtil.addCondition(lsSQL, "a.sStockIDx = " + SQLUtil.toSQL(lsFilter));
-        lsSQL = MiscUtil.addCondition(lsSQL, fsCondition);
-        
-        System.out.println(lsSQL);
-
+        lsSQL = MiscUtil.addCondition(lsSQL, "a.sStockIDx = " + SQLUtil.toSQL(fsCondition));
+        lsSQL = MiscUtil.addCondition(lsSQL, "a.nEntryNox = " + SQLUtil.toSQL(fnEntryNox));
+        System.out.print("this is lsSQL openrec WithCondition == " + lsSQL + "\n");
         ResultSet loRS = poGRider.executeQuery(lsSQL);
-
         try {
             if (loRS.next()) {
                 for (int lnCtr = 1; lnCtr <= loRS.getMetaData().getColumnCount(); lnCtr++) {
@@ -270,7 +281,8 @@ public class Model_Inv_Ledger implements GEntity {
                 }
 
                 pnEditMode = EditMode.UPDATE;
-
+                
+                
                 poJSON.put("result", "success");
                 poJSON.put("message", "Record loaded successfully.");
             } else {
@@ -285,7 +297,6 @@ public class Model_Inv_Ledger implements GEntity {
         return poJSON;
     }
 
-
     /**
      * Save the entity.
      *
@@ -299,10 +310,9 @@ public class Model_Inv_Ledger implements GEntity {
             String lsSQL;
             if (pnEditMode == EditMode.ADDNEW) {
                 //replace with the primary key column info
-//                setStockID(MiscUtil.getNextCode(getTable(), "sStockIDx", false, poGRider.getConnection(), ""));
+//                setStockID(MiscUtil.getNextCode(getTable(), "a.sStockIDx", true, poGRider.getConnection(), poGRider.getBranchCode()));
 
                 setModifiedDate(poGRider.getServerDate());
-                setModifiedBy(poGRider.getUserID());
                 lsSQL = makeSQL();
 
                 if (!lsSQL.isEmpty()) {
@@ -318,17 +328,16 @@ public class Model_Inv_Ledger implements GEntity {
                     poJSON.put("message", "No record to save.");
                 }
             } else {
-                Model_Inv_Ledger loOldEntity = new Model_Inv_Ledger(poGRider);
+                Model_Inventory_Sub_Unit loOldEntity = new Model_Inventory_Sub_Unit(poGRider);
+                
+                //replace with the primary key column info
+                JSONObject loJSON = loOldEntity.openRecordWithCondition(this.getStockID(), this.getEntryNox());
 
                 setModifiedDate(poGRider.getServerDate());
-                setModifiedBy(poGRider.getUserID());
-                //replace with the primary key column info
-                JSONObject loJSON = loOldEntity.openRecord(this.getStockID());
-
                 if ("success".equals((String) loJSON.get("result"))) {
                     //replace the condition based on the primary key column of the record
-                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sStockIDx = " + SQLUtil.toSQL(this.getStockID()),  "xBarCodex»xDescript»xWHouseNm»xBranchNm");
-
+                    lsSQL = MiscUtil.makeSQL(this, loOldEntity, "sStockIDx = " + SQLUtil.toSQL(this.getStockID()) + " AND nEntryNox = " + SQLUtil.toSQL(this.getEntryNox()),  "xBarCodex»xDescript»xBarCodeU»xDescripU»xMeasurID»xMeasurNm");
+                    System.out.println("Update lsSQL value == " + lsSQL);
                     if (!lsSQL.isEmpty()) {
                         if (poGRider.executeQuery(lsSQL, getTable(), poGRider.getBranchCode(), "") > 0) {
                             poJSON.put("result", "success");
@@ -410,276 +419,26 @@ public class Model_Inv_Ledger implements GEntity {
         return (String) getValue("sStockIDx");
     }
 
+
     /**
-     * @return The sBranchCd of this record.
-     */
-    public String getBranchCode() {
-        return (String) getValue("sBranchCd");
-    }
-    
-    /**
-     * Sets the sBrandCde of this record.
+     * Sets the sStockIDx of this record.
      *
      * @param fsValue
      * @return result as success/failed
      */
-    public JSONObject setBranchCode(String fsValue) {
-        return setValue("sBranchCd", fsValue);
-    }
-    
-  
-    /**
-     * @return The sWHouseID of this record.
-     */
-    public String getWHouseID() {
-        return (String) getValue("sWHouseID");
-    }
-    
-    /**
-     * Sets the sWHouseID of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setWHouseID(String fsValue) {
-        return setValue("sWHouseID", fsValue);
-    }
-    
-    /**
-     * @return The nLedgerNo of this record.
-     */
-    public int getLedgerNo() {
-        return (int) getValue("nLedgerNo");
-    }
-    
-    /**
-     * Sets the nLedgerNo of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setLedgerNo(Number fsValue) {
-        return setValue("nLedgerNo", fsValue);
+    public JSONObject setSubItemID(String fsValue) {
+        return setValue("sItmSubID", fsValue);
     }
 
     /**
-     * Sets the date and time the record was transact.
-     *
-     * @param fdValue
-     * @return result as success/failed
+     * @return The sStockIDx of this record.
      */
-    public JSONObject setTransactDate(Date fdValue) {
-        return setValue("dTransact", fdValue);
+    public String getSubItemID() {
+        return (String) getValue("sItmSubID");
     }
 
     /**
-     * @return The date and time the record was transact.
-     */
-    public Date getTransactDate() {
-        return (Date) getValue("dTransact");
-    }
-
-    
-    /**
-     * @return The sSourceCd of this record.
-     */
-    public String getSourceCode() {
-        return (String) getValue("sSourceCd");
-    }
-    
-    /**
-     * Sets the sSourceCd of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setSourceCode(String fsValue) {
-        return setValue("sSourceCd", fsValue);
-    }
-    
-    /**
-     * @return The sSourceNo of this record.
-     */
-    public String getSourceNo() {
-        return (String) getValue("sSourceNo");
-    }
-    
-    /**
-     * Sets the sSourceNo of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setSourceNo(String fsValue) {
-        return setValue("sSourceNo", fsValue);
-    }
-
-    /**
-    /**
-     * Sets the nQtyInxxx .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityIn(Number fsValue){
-        return setValue("nQtyInxxx", fsValue);
-    }
-    
-    /**
-     * @return The nQtyInxxx. 
-     */
-    public Object getQuantityIn(){
-        return (Object) getValue("nQtyInxxx");
-    }
-    
-    /**
-    /**
-     * Sets the nQtyOutxx .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityOut(Number fsValue){
-        return setValue("nQtyOutxx", fsValue);
-    }
-    
-    /**
-     * @return The nQtyOutxx. 
-     */
-    public Object getQuantityOut(){
-        return (Object) getValue("nQtyOutxx");
-    }
-    
-
-    /**
-    /**
-     * Sets the nQtyOrder .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityOrder(Number fsValue){
-        return setValue("nQtyOrder", fsValue);
-    }
-    
-    /**
-     * @return The nQtyOrder. 
-     */
-    public Object getQuantityOrder(){
-        return (Object) getValue("nQtyOrder");
-    }
-
-    /**
-    /**
-     * Sets the nQtyIssue .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityIssue(Number fsValue){
-        return setValue("nQtyIssue", fsValue);
-    }
-    
-    /**
-     * @return The nQtyIssue. 
-     */
-    public Object getQuantityIssue(){
-        return (Object) getValue("nQtyIssue");
-    }
-    
-    /**
-    /**
-     * Sets the nPurPrice .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setPurchasePrice(Number fsValue){
-        return setValue("nPurPrice", fsValue);
-    }
-    
-    /**
-     * @return The nUnitPrce. 
-     */
-    public Object getPurchasePrice(){
-        return (Object) getValue("nPurPrice");
-    }
-    
-    /**
-    /**
-     * Sets the nUnitPrce .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setUnitPrice(Number fsValue){
-        return setValue("nUnitPrce", fsValue);
-    }
-    
-    /**
-     * @return The nUnitPrce. 
-     */
-    public Object getUnitPrice(){
-        return (Object) getValue("nUnitPrce");
-    }
-    
-    
-    /**
-     * Sets the nSelPrice .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setSelPrice(Number fsValue){
-        return setValue("nSelPrice", fsValue);
-    }
-    
-    /**
-     * @return The nSelPrice. 
-     */
-    public Object getSelPrice(){
-        return (Object) getValue("nSelPrice");
-    }
-    
-    
-    /**
-    /**
-     * Sets the nQtyOnHnd .
-     * 
-     * @param fsValue 
-     * @return  True if the record assignment is successful.
-     */
-    public JSONObject setQuantityOnHand(Number fsValue){
-        return setValue("nQtyOnHnd", fsValue);
-    }
-    
-    /**
-     * @return The nQtyOnHnd. 
-     */
-    public Object getQuantityOnHand(){
-        return (Object) getValue("nQtyOnHnd");
-    }
-    
-    
-    /**
-     * Sets the date and time the record expiration date.
-     *
-     * @param fdValue
-     * @return result as success/failed
-     */
-    public JSONObject setExpiryDate(Date fdValue) {
-        return setValue("dExpiryxx", fdValue);
-    }
-
-    /**
-     * @return The date and time the record expiration date.
-     */
-    public Date getExpiryDate() {
-        return (Date) getValue("dExpiryxx");
-    } 
-    
-    /**
-     * Sets the xBarCodex of this record.
+     * Sets the sBarCodex of this record.
      *
      * @param fsValue
      * @return result as success/failed
@@ -689,96 +448,80 @@ public class Model_Inv_Ledger implements GEntity {
     }
 
     /**
-     * @return The xBarCodex of this record.
+     * @return The sBarCodex of this record.
      */
     public String getBarcode() {
         return (String) getValue("xBarCodex");
     }
-    
+
     /**
-     * Sets the xDescript of this record.
+     * Sets the sDescript Code of this record.
      *
      * @param fsValue
      * @return result as success/failed
      */
-    public JSONObject setDescript(String fsValue) {
+    public JSONObject setDescription(String fsValue) {
         return setValue("xDescript", fsValue);
     }
 
     /**
-     * @return The xDescript of this record.
+     * @return The sDescript Code of this record.
      */
-    public String getDescript() {
+    public String getDescription() {
         return (String) getValue("xDescript");
     }
+    
+
     /**
-     * Sets the xWHouseNm of this record.
+     * Sets the sBriefDsc Code of this record.
      *
      * @param fsValue
      * @return result as success/failed
      */
-    public JSONObject setWareHouseName(String fsValue) {
-        return setValue("xWHouseNm", fsValue);
+    public JSONObject setSubItemDescription(String fsValue) {
+        return setValue("xDescripU", fsValue);
     }
 
     /**
-     * @return The xWHouseNm of this record.
+     * @return The sBriefDsc Code of this record.
      */
-    public String getWareHouseName() {
-        return (String) getValue("xWHouseNm");
-    }
-
-    /**
-     * Sets the xBranchNm of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
-     */
-    public JSONObject setBranchName(String fsValue) {
-        return setValue("xBranchNm", fsValue);
-    }
-
-    /**
-     * @return The xBranchNm of this record.
-     */
-    public String getBranchName() {
-        return (String) getValue("xBranchNm");
+    public String getSubItemDescription() {
+        return (String) getValue("xDescripU");
     }
     
     /**
-     * Sets the Inventory RecdStat of this record.
-     *
-     * @param fsValue
-     * @return result as success/failed
+     * @return The nQuantity. 
      */
-    public JSONObject setRecdStat(String fsValue) {
-        return setValue("cRecdStat", fsValue);
+    public Number getQuantity(){
+        return (Number) getValue("nQuantity");
     }
-
+    
     /**
-     * @return The Category RecdStat of this record.
+     * Sets the nQuantity .
+     * 
+     * @param fsValue 
+     * @return  True if the record assignment is successful.
      */
-    public String getRecdStat() {
-        return (String) getValue("cRecdStat");
+    public JSONObject setQuantity(Number fsValue){
+        return setValue("nQuantity", fsValue);
     }
-
     /**
-     * Sets record as active.
-     *
-     * @param fbValue
-     * @return result as success/failed
+     * @return The nEntryNox. 
      */
-    public JSONObject setActive(boolean fbValue) {
-        return setValue("cRecdStat", fbValue ? "1" : "0");
+    public Integer getEntryNox(){
+        return (Integer) getValue("nEntryNox");
     }
-
+    
     /**
-     * @return If record is active.
+     * Sets the nEntryNox .
+     * 
+     * @param fsValue 
+     * @return  True if the record assignment is successful.
      */
-    public boolean isActive() {
-        return ((String) getValue("cRecdStat")).equals("1");
+    public JSONObject setEntryNox(Number fsValue){
+        return setValue("nEntryNox", fsValue);
     }
-
+    
     /**
      * Sets the user encoded/updated the record.
      *
@@ -812,33 +555,39 @@ public class Model_Inv_Ledger implements GEntity {
     public Date getModifiedDate() {
         return (Date) getValue("dModified");
     }
-    public String getSQL(){
-        return "SELECT" +
-                        "   a.sStockIDx" +
-                        " , a.sBranchCd" +
-                        " , a.sWHouseID" +
-                        " , a.nLedgerNo" +
-                        " , a.dTransact" +
-                        " , a.sSourceCd" +
-                        " , a.sSourceNo" +
-                        " , a.nQtyInxxx" +
-                        " , a.nQtyOutxx" +
-                        " , a.nQtyOrder" +
-                        " , a.nQtyIssue" +
-                        " , a.nPurPrice" +
-                        " , a.nUnitPrce" +
-                        " , a.nQtyOnHnd" +
-                        " , a.dExpiryxx" +
-                        " , a.sModified" +
-                        " , a.dModified" +
-                        " , b.sBarCodex xBarCodex" +
-                        " , b.sDescript xDescript" +
-                        " , c.sWHouseNm xWHouseNm" +
-                        " , d.sBranchNm xBranchNm" +
-                        " FROM Inv_Ledger a" +
-                        "    LEFT JOIN Inventory b ON a.sStockIDx = b.sStockIDx" +
-                        "    LEFT JOIN Warehouse c ON a.sWhouseID = c.sWhouseID" +
-                        "    LEFT JOIN Branch d ON a.sBranchCd = d.sBranchCd";
+
+    /**
+     * @return The sMeasurID of this record.
+     */
+    public String getMeasureID() {
+        return (String) getValue("sMeasurID");
+    }
+    
+    /**
+     * Sets the sMeasurID of this record.
+     *
+     * @param fsValue
+     * @return result as success/failed
+     */
+    public JSONObject setMeasureID(String fsValue) {
+        return setValue("sMeasurID", fsValue);
+    }
+    
+    /**
+     * @return The xMeasurNm of this record.
+     */
+    public String getMeasureName() {
+        return (String) getValue("xMeasurNm");
+    }
+    
+    /**
+     * Sets the xMeasurNm of this record.
+     *
+     * @param fsValue
+     * @return result as success/failed
+     */
+    public JSONObject setMeasureName(String fsValue) {
+        return setValue("xMeasurNm", fsValue);
     }
     /**
      * Gets the SQL statement for this entity.
@@ -846,7 +595,7 @@ public class Model_Inv_Ledger implements GEntity {
      * @return SQL Statement
      */
     public String makeSQL() {
-        return MiscUtil.makeSQL(this, "xBarCodex»xDescript»xWHouseNm»xBranchNm");
+        return MiscUtil.makeSQL(this, "xBarCodex»xDescript»xBarCodeU»xDescripU»xMeasurID»xMeasurNm");
     }
 
     /**
@@ -855,9 +604,27 @@ public class Model_Inv_Ledger implements GEntity {
      * @return SelectSQL Statement
      */
     public String makeSelectSQL() {
-        return MiscUtil.makeSelect(this, "xBarCodex»xDescript»xWHouseNm»xBranchNm");
+        return MiscUtil.makeSelect(this, "xBarCodex»xDescript»xBarCodeU»xDescripU»xMeasurID»xMeasurNm");
     }
 
+    public String getSQL(){
+        return "SELECT" +
+                "  a.sStockIDx" +
+                ", a.nEntryNox" +
+                ", a.sItmSubID" +
+                ", a.nQuantity" +
+                ", a.dModified" +
+                ", b.sBarCodex xBarCodex" +
+                ", b.sDescript xDescript" +
+                ", c.sBarCodex xBarCodeU" +
+                ", c.sDescript xDescripU" +
+                ", c.sMeasurID xMeasurID" +
+                ", d.sMeasurNm xMeasurNm" +
+            " FROM Inventory_Sub_Unit a" +
+                " LEFT JOIN Inventory b ON a.sStockIDx = b.sStockIDx" +
+                " LEFT JOIN Inventory c ON a.sItmSubID = c.sStockIDx" + 
+                " LEFT JOIN Measure d ON c.sMeasurID = d.sMeasurID";
+    }
     private void initialize() {
         try {
             poEntity = MiscUtil.xml2ResultSet(System.getProperty("sys.default.path.metadata") + XML, getTable());
@@ -866,7 +633,8 @@ public class Model_Inv_Ledger implements GEntity {
             poEntity.moveToInsertRow();
 
             MiscUtil.initRowSet(poEntity);
-//            poEntity.updateString("cRecdStat", RecordStatus.ACTIVE);
+            poEntity.updateObject("nEntryNox", 1);
+            poEntity.updateObject("nQuantity", 0.0);
 
             poEntity.insertRow();
             poEntity.moveToCurrentRow();
