--- src/main/java/org/guanzon/cas/model/inventory/Inv_Master.java
@@ -1,4 +1,0 @@
-package org.guanzon.cas.model.inventory;
-
-public class Inv_Master {
-}
