--- src/main/java/org/guanzon/cas/model/inventory/Inventory.java
@@ -1,4 +1,0 @@
-package org.guanzon.cas.model.inventory;
-
-public class Inventory {
-}
